refactored, MesherSteps separated out.
